% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{get_hud_fmr_listcounties}
\alias{get_hud_fmr_listcounties}
\title{Get HUD List of County Codes data.}
\usage{
get_hud_fmr_listcounties(stateid = "AL", hud_key)
}
\arguments{
\item{stateid}{Character string indicating the state.}

\item{hud_key}{Character string indicating your API key from HUD.}
}
\value{
A data table containing a list of county codes from HUD.
}
\description{
Provides list of county codes data from HUD.
}
\details{
Provides list of county codes data from HUD.
}
\examples{
\dontrun{
library(hudr)

hud_key <- Sys.getenv("HUD_API_KEY")

state_codes <- get_hud_fmr_listcounties(hud_key = hud_key)
}
}
\author{
Paul Richardson
}
