% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{get_hud_fmr_listmetros}
\alias{get_hud_fmr_listmetros}
\title{Get HUD List of MSA Codes data.}
\usage{
get_hud_fmr_listmetros(hud_key)
}
\arguments{
\item{hud_key}{Character string indicating your API key from HUD.}
}
\value{
A data table containing a list of MSA codes from HUD.
}
\description{
Provides list of MSA codes data from HUD.
}
\details{
Provides list of MSA codes from HUD.
}
\examples{
\dontrun{
library(hudr)

hud_key <- Sys.getenv("HUD_API_KEY")

cbsa_codes <- get_hud_fmr_listmetros(hud_key = hud_key)
}
}
\author{
Paul Richardson
}
