% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr.R
\name{get_hud_fmr_liststates}
\alias{get_hud_fmr_liststates}
\title{Get HUD List of State Codes data.}
\usage{
get_hud_fmr_liststates(hud_key)
}
\arguments{
\item{hud_key}{Character string indicating your API key from HUD.}
}
\value{
A data table containing a list of state codes from HUD.
}
\description{
Provides list of state codes data from HUD.
}
\details{
Provides list of state codes data from HUD.
}
\examples{
\dontrun{
library(hudr)

hud_key <- Sys.getenv("HUD_API_KEY")

state_codes <- get_hud_fmr_liststates(hud_key = hud_key)
}
}
\author{
Paul Richardson
}
