% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudr-package.R
\docType{package}
\name{hudr-package}
\alias{hudr}
\alias{hudr-package}
\title{Package that provides data from HUD.}
\description{
Provides functions to access data from the US Department of Housing and Urban Development <https://www.huduser.gov/portal/dataset/fmr-api.html>.
}
\details{
Provides data from the US Department of Housing and Urban Development (HUD).
}
\references{
<https://www.huduser.gov/portal/dataset/fmr-api.html>
}
